+++ frontend/admin/js/a_dashboard.js
@@ -1,0 +1,1 @@
+import showToast from "./a_toast.js";